+++ src/modules/settings/index.js
@@ -2,11 +2,14 @@ import React from 'react';
 import { Route, Redirect, Switch } from 'react-router-dom';
 import { Segment } from 'semantic-ui-react';
 import { SemanticToastContainer } from 'react-semantic-toasts';
+import { connect } from 'react-redux';
 import 'react-semantic-toasts/styles/react-semantic-alert.css';
 
 import Security from './Security';
 import General from './General';
 import NavSwitcher from '../shared/components/NavSwitcher';
+import Loader from '../shared/components/Loader';
+import { getLoadingState } from './selectors';
 
 const navSwitcherConfig = [
   { name: 'Settings' },
@@ -14,13 +17,14 @@ const navSwitcherConfig = [
   { name: 'Security', goTo: '/settings/security' },
 ];
 
-const Settings = () => (
+const Settings = ({ isLoading }) => (
   <>
+    {isLoading && <Loader />}
     <Segment>
-      <SemanticToastContainer position='top-right' />
+      <SemanticToastContainer position="top-right" />
       <NavSwitcher config={navSwitcherConfig} />
       <Switch>
-        <Route exact path="/settings/general" component={General}  />
+        <Route exact path="/settings/general" component={General} />
         <Route exact path="/settings/security" component={Security} />
         <Redirect to="/settings/security" />
       </Switch>
@@ -28,4 +32,8 @@ const Settings = () => (
   </>
 );
 
-export default Settings;
+const mapStateToProps = state => ({
+  isLoading: getLoadingState(state),
+});
+
+export default connect(mapStateToProps)(Settings);